Lab 08/09 - Database Refactoring
Realized that I had used the object keys, rather than the API keys,
for the DB column names. So I had to do an ugly thing last night
where I was creating a new object that I was passing to the Location
constructor when building a new object from cached data. Fixed it
so that DB columns now matach the API's keys.